Improvements

+++ src/root/parser/parse_function/parse_literal.rs
@@ -17,7 +17,7 @@ pub struct LiteralToken {
 #[derive(Debug)]
 pub enum LiteralTokens {
     Bool(bool),
-    Int(i64),
+    Int(i128),
 }
 
 impl LiteralTokens {
@@ -39,7 +39,7 @@ pub fn parse_literal(s: Span) -> ParseResult<Span, LiteralToken> {
     let (ns, l) = alt((
         |x| tag("true")(x).map(|(s, _)| (s, LiteralTokens::Bool(true))),
         |x| tag("false")(x).map(|(s, _)| (s, LiteralTokens::Bool(false))),
-        |x| nom::character::complete::i64(x).map(|(s, i)| (s, LiteralTokens::Int(i))),
+        |x| nom::character::complete::i128(x).map(|(s, i)| (s, LiteralTokens::Int(i))),
     ))(s)?;
 
     let l = LiteralToken {
